Remove no more used classes (since transport layer refactoring is done)
