--- leshan-core-cf/src/main/java/org/eclipse/leshan/core/californium/EndpointFactory.java
@@ -1,39 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2017 Sierra Wireless and others.
- *
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v2.0
- * and Eclipse Distribution License v1.0 which accompany this distribution.
- *
- * The Eclipse Public License is available at
- *    http://www.eclipse.org/legal/epl-v20.html
- * and the Eclipse Distribution License is available at
- *    http://www.eclipse.org/org/documents/edl-v10.html.
- *
- * Contributors:
- *     Sierra Wireless - initial API and implementation
- *     Rikard Höglund (RISE SICS) - Additions to support OSCORE
- *******************************************************************************/
-package org.eclipse.leshan.core.californium;
-
-import java.net.InetSocketAddress;
-
-import org.eclipse.californium.core.network.CoapEndpoint;
-import org.eclipse.californium.core.observe.ObservationStore;
-import org.eclipse.californium.elements.config.Configuration;
-import org.eclipse.californium.oscore.OSCoreCtxDB;
-import org.eclipse.californium.scandium.config.DtlsConnectorConfig;
-
-/**
- * Class used to delegate CoAP endpoint creation in all Leshan Builders.
- *
- * @see DefaultEndpointFactory
- */
-public interface EndpointFactory {
-
-    CoapEndpoint createUnsecuredEndpoint(InetSocketAddress address, Configuration coapConfig, ObservationStore store,
-            OSCoreCtxDB db);
-
-    CoapEndpoint createSecuredEndpoint(DtlsConnectorConfig dtlsConfig, Configuration coapConfig, ObservationStore store,
-            OSCoreCtxDB db);
-}
