+++ leshan-core-cf/src/main/java/org/eclipse/leshan/core/californium/SyncRequestObserver.java
@@ -33,10 +33,6 @@ import org.eclipse.leshan.core.response.LwM2mResponse;
  */
 public abstract class SyncRequestObserver<T extends LwM2mResponse> extends CoapSyncRequestObserver {
 
-    public SyncRequestObserver(Request coapRequest, long timeout) {
-        this(coapRequest, timeout, new TemporaryExceptionTranslator());
-    }
-
     public SyncRequestObserver(Request coapRequest, long timeout, ExceptionTranslator exceptionTranslator) {
         super(coapRequest, timeout, exceptionTranslator);
     }
